hwcomposer: Move wayland_thread to struct display
